reorganize: move tests to own folder
