+++ .gitignore
@@ -1,6 +1,6 @@
 # File created using '.gitignore Generator' for Visual Studio Code: https://bit.ly/vscode-gig
-# Created by https://www.toptal.com/developers/gitignore/api/visualstudiocode,macos,dotenv,linux,react
-# Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,macos,dotenv,linux,react
+# Created by https://www.toptal.com/developers/gitignore/api/visualstudiocode,react,macos,linux,dotenv,storybookjs
+# Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,react,macos,linux,dotenv,storybookjs
 
 ### dotenv ###
 .env
@@ -69,6 +69,12 @@ psd
 thumb
 sketch
 
+### StorybookJs ###
+# gitignore template for the Storybook, UI guide for front apps
+# website: https://storybook.js.org/
+
+storybook-static/
+
 ### VisualStudioCode ###
 .vscode/*
 !.vscode/settings.json
@@ -88,10 +94,13 @@ sketch
 .history
 .ionide
 
-# End of https://www.toptal.com/developers/gitignore/api/visualstudiocode,macos,dotenv,linux,react
+# End of https://www.toptal.com/developers/gitignore/api/visualstudiocode,react,macos,linux,dotenv,storybookjs
 
 # Custom rules (everything added below won't be overriden by 'Generate .gitignore File' if you use 'Update' option)
+
 .env.*
 !.env.template
 .vinxi
 .output
+*storybook.log
+coverage/
