+++ package.json
@@ -19,7 +19,9 @@
         "fix:format": "biome format --write ./",
         "fix:lint": "biome lint --write ./",
         "test": "vitest run",
-        "test:watch": "vitest"
+        "test:watch": "vitest",
+        "storybook": "storybook dev -p 6006",
+        "build-storybook": "storybook build"
     },
     "keywords": [],
     "author": "",
@@ -51,15 +53,26 @@
     },
     "devDependencies": {
         "@biomejs/biome": "^1.9.3",
+        "@chromatic-com/storybook": "^1.9.0",
+        "@storybook/addon-essentials": "^8.3.6",
+        "@storybook/addon-interactions": "^8.3.6",
+        "@storybook/addon-links": "^8.3.6",
+        "@storybook/addon-onboarding": "^8.3.6",
+        "@storybook/blocks": "^8.3.6",
+        "@storybook/react": "^8.3.6",
+        "@storybook/react-vite": "^8.3.6",
+        "@storybook/test": "^8.3.6",
         "@testing-library/jest-dom": "^6.5.0",
         "@testing-library/react": "^16.0.1",
         "@types/react": "^18.3.11",
         "@types/react-dom": "^18.3.0",
         "@vitejs/plugin-react": "^4.3.2",
+        "@vitest/coverage-istanbul": "^2.1.2",
         "autoprefixer": "^10.4.20",
         "husky": "^9.1.6",
         "jsdom": "^25.0.1",
         "postcss": "^8.4.47",
+        "storybook": "^8.3.6",
         "tailwindcss": "^3.4.13",
         "typescript": "^5.6.2",
         "vite-tsconfig-paths": "^5.0.1",
