+++ tsconfig.json
@@ -21,5 +21,10 @@
         "noUncheckedIndexedAccess": true,
         "noEmit": true
     },
-    "include": ["app/**/*.ts", "app/**/*.tsx"]
+    "include": [
+        "app/**/*.ts",
+        "app/**/*.tsx",
+        "testing/**/*.ts",
+        "testing/**/*.tsx"
+    ]
 }
